# Add specializations for istriu/istril to speed up isdiag.